+++ R/DSC_streamkm.R
@@ -20,46 +20,60 @@
 #'
 #' This is an interface to the MOA implementation of streamKM++.
 #'
-#' streamKM++ uses a tree-based sampling strategy to obtain a small weighted sample of the stream called coreset.
-#' Upon reclustering, the algorithm applies the k-means++ algorithm to find a given number of centres in the coreset.
+#' streamKM++ uses a tree-based sampling strategy to obtain a small weighted sample of the stream
+#' called coreset. The MOA implementation applies the k-means++ algorithm to find a given number
+#' of centers in the coreset.
+#'
+#' **Notes:**
+#'
+#'  * The clustere can only cluster the number of points specified in `length` ans then
+#'    produces an `ArrayIndexOutOfBoundsException` error.
+#'  * The coreset (micro-clusters are not accessible), only the macro-clusters can be requested.
+#'
+#' @family DSC_MOA
 #'
 #' @param sizeCoreset Size of the coreset
 #' @param numClusters Number of clusters to compute
 #' @param length Length of the data stream
+#' @param ... Further arguments are passed on to [DSC_Kmeans] for reclustering.
 #'
 #' @aliases DSC_StreamKM streamkm StreamKM
 #' @author Matthias Carnein
 #'
-#' @references Marcel R. Ackermann, Christiane Lammersen, Marcus Maertens, Christoph Raupach, Christian Sohler, Kamil Swierkot. "StreamKM++: A Clustering Algorithm for Data Streams." In: Proceedings of the 12th Workshop on Algorithm Engineering and Experiments (ALENEX '10), 2010
-#'
+#' @references
+#' Marcel R. Ackermann, Christiane Lammersen, Marcus Maertens, Christoph Raupach,
+#' Christian Sohler, Kamil Swierkot.
+#' StreamKM++: A Clustering Algorithm for Data Streams. In: _Proceedings of the 12th Workshop on Algorithm
+#' Engineering and Experiments (ALENEX '10)_, 2010.
 #' @examples
-#' # data with 3 clusters
-#' stream <- DSD_Gaussians(k=3, d=2)
+#' set.seed(1000)
+#' stream <- DSD_Gaussians(k = 3, d = 2, noise = 0.05)
 #'
 #' # cluster with streamKM++
-#' streamkm <- DSC_StreamKM(sizeCoreset=10000, numClusters=3, length=10000)
-#' update(streamkm, stream, 10000)
+#' streamkm <- DSC_StreamKM(sizeCoreset = 100, numClusters = 3, length = 1000)
+#' update(streamkm, stream, 100)
 #' streamkm
 #'
-#' # plot macro-clusters
-#' plot(streamkm, stream, type="macro")
-#'
-#' @export DSC_StreamKM
-DSC_StreamKM <- function(sizeCoreset=10000, numClusters=5, length=100000L) {
+#' # plot macro-clusters (no access to micro-clusters)
+#' plot(streamkm, stream)
+#' @export
+DSC_StreamKM <-
+  function(sizeCoreset = 10000,
+    numClusters = 5,
+    length = 100000L,
+    ...) {
+    ### Java code does parameter checking
+    # streakm options:
+    # -s sizeCoreset 	10000 (Size of the coreset (m))
+    # -k numClusters 	5 (Number of clusters to compute)
+    # -l length 100000 (Length of the data stream (n), range: 0, Max_int)
+    paramList <- list(
+      s = as.integer(sizeCoreset),
+      k = as.integer(numClusters),
+      l = as.integer(length)
+    )
 
-  ### Java code does parameter checking
-  # streakm options:
-  # -s sizeCoreset 	10000 (Size of the coreset (m))
-  # -k numClusters 	5 (Number of clusters to compute)
-  # -l length 100000 (Length of the data stream (n), range: 0, Max_int)
-  paramList <- list(
-    s=as.integer(sizeCoreset),
-    k=as.integer(numClusters),
-    l=as.integer(length)
-  )
+    ## MOA implementation does not return micro-clusters but only macro clusters
+    DSC_MOA_Clusterer("moa/clusterers/streamkm/StreamKM", "StreamKM", paramList)
 
-  ## MOA implementation does not return micro-clusters but only macro clusters
-  clus <- DSC_MOA_Clusterer("moa/clusterers/streamkm/StreamKM", "StreamKM", paramList)
-
-  clus
-}
+  }